enhancement(bys-dispatch): add background dispatch
