+++ server/graphql/v1/mutations/backyourstack.js
@@ -1,6 +1,6 @@
 import moment from 'moment';
 import models from '../../../models';
-import { dispatchFunds, getNextDispatchingDate, needsDispatching } from '../../../lib/backyourstack/dispatcher';
+import { backgroundDispatch, needsDispatching } from '../../../lib/backyourstack/dispatcher';
 
 import status from '../../../constants/order_status';
 
@@ -25,23 +25,7 @@ export async function dispatchOrder(orderId) {
       `Your BackYourStack order is already complete for this month. The next dispatch of funds will be on ${nextDispatchDate}`,
     );
   }
-
-  let dispatchedOrders;
-  try {
-    dispatchedOrders = await dispatchFunds(order);
-  } catch (err) {
-    console.error(err);
-    throw new Error(`Unable to dispatch funds to collectves.`);
-  }
-
-  if (dispatchedOrders) {
-    const currentDispatchDate = (subscription.data && subscription.data.nextDispatchDate) || new Date();
-    subscription.data = {
-      nextDispatchDate: getNextDispatchingDate(subscription.interval, currentDispatchDate),
-    };
-
-    await subscription.save();
-  }
-
-  return dispatchedOrders;
+  // Funds are dispatched asynchronously
+  backgroundDispatch(order, subscription);
+  return { dispatching: true };
 }
